LayoutTests/fast/speechsynthesis/voices-non-mock-expected.txt

This tests that we can get synthesizer voices.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS voiceCount > 20 is true
PASS foundEnglishVoice is true
PASS foundDefaultVoice is true
PASS successfullyParsed is true

TEST COMPLETE
